removing combined analysis
