+++ autofit/non_linear/search/abstract_search.py
@@ -46,8 +46,6 @@ from autofit.non_linear.samples.samples import Samples
 from autofit.non_linear.samples.summary import SamplesSummary
 from autofit.non_linear.timer import Timer
 from autofit.non_linear.analysis import Analysis
-from autofit.non_linear.analysis.combined import CombinedResult
-from autofit.non_linear.analysis.indexed import IndexCollectionAnalysis
 from autofit.non_linear.paths.null import NullPaths
 from autofit.graphical.declarative.abstract import PriorFactor
 from autofit.graphical.expectation_propagation import AbstractFactorOptimiser
@@ -467,67 +465,6 @@ class NonLinearSearch(AbstractFactorOptimiser, ABC):
 
         except ModuleNotFoundError:
             return False
-
-    def fit_sequential(
-        self,
-        model: AbstractPriorModel,
-        analysis: IndexCollectionAnalysis,
-        info: Optional[Dict] = None,
-    ) -> CombinedResult:
-        """
-        Fit multiple analyses contained within the analysis sequentially.
-
-        This can be useful for avoiding very high dimensional parameter spaces.
-
-        Parameters
-        ----------
-        analysis
-            Multiple analyses that are fit sequentially
-        model
-            An object that represents possible instances of some model with a
-            given dimensionality which is the number of free dimensions of the
-            model.
-        info
-            Optional dictionary containing information about the fit that can be loaded by the aggregator.
-
-        Returns
-        -------
-        An object combining the results of each individual optimisation.
-
-        Raises
-        ------
-        AssertionError
-            If the model has 0 dimensions.
-        ValueError
-            If the analysis is not a combined analysis
-        """
-        results = []
-
-        _paths = self.paths
-        original_name = self.paths.name or "analysis"
-
-        model = analysis.modify_model(model=model)
-
-        try:
-            if not isinstance(model, Collection):
-                model = [model for _ in range(len(analysis.analyses))]
-        except AttributeError:
-            raise ValueError(
-                f"Analysis with type {type(analysis)} is not supported by fit_sequential"
-            )
-
-        for i, (model, analysis) in enumerate(zip(model, analysis.analyses)):
-            self.paths = copy.copy(_paths)
-            self.paths.name = f"{original_name}/{i}"
-            results.append(
-                self.fit(
-                    model=model,
-                    analysis=analysis,
-                    info=info,
-                )
-            )
-        self.paths = _paths
-        return CombinedResult(results)
 
     def fit(
         self,
